+++ workspaces/jfrog-artifactory/plugins/jfrog-artifactory/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @backstage-community/plugin-jfrog-artifactory
 
+## 1.26.0
+
+### Minor Changes
+
+- 1c3b8f4: Backstage version bump to v1.47.2
+
 ## 1.25.2
 
 ### Patch Changes